refactor: to separate styles, and reorganise components
