+++ src/Header.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
-import { View, Text } from "react-native";
+import { View, Text, StyleSheet } from "react-native";
 import { colors } from "../colors";
 
 export const Header: FC<{
@@ -11,41 +11,46 @@ export const Header: FC<{
 
   return (
     <View
-      style={{
-        paddingTop: top,
-        width: "100%",
-        backgroundColor: colors.gray["600"],
-      }}
+      style={[
+        styles.container,
+        {
+          paddingTop: top,
+        },
+      ]}
     >
       <View
-        style={{
-          height: headerHeight,
-          width: "100%",
-          alignItems: "center",
-          justifyContent: "center",
-        }}
-      >
-        <Text
-          style={[
-            {
-              fontSize: 17,
-              fontWeight: "600",
-              letterSpacing: -0.41,
-            },
-          ]}
-        >
-          Settings
-        </Text>
-      </View>
-      <View
         style={[
+          styles.textContainer,
           {
-            width: "100%",
-            height: 1,
-            backgroundColor: "#ddd",
+            height: headerHeight,
           },
         ]}
-      />
+      >
+        <Text style={styles.text}>Settings</Text>
+      </View>
+      <View style={styles.border} />
     </View>
   );
 };
+
+const styles = StyleSheet.create({
+  container: {
+    width: "100%",
+    backgroundColor: colors.gray["600"],
+  },
+  border: {
+    width: "100%",
+    height: 1,
+    backgroundColor: "#ddd",
+  },
+  text: {
+    fontSize: 17,
+    fontWeight: "600",
+    letterSpacing: -0.41,
+  },
+  textContainer: {
+    width: "100%",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+});
