+++ src/Search.tsx
@@ -1,44 +1,44 @@
 import { FC } from "react";
-import {Image, TextInput, View} from "react-native";
-
+import { Image, StyleSheet, TextInput, View } from "react-native";
 
 export const Search: FC<{
   searchHeight: number;
 }> = ({ searchHeight }) => {
-
   return (
     <View
       style={[
+        styles.container,
         {
-          flex: 1,
           height: searchHeight,
-          borderRadius: 10,
-          backgroundColor: "#E3E3E8",
-          marginHorizontal: 16,
         },
       ]}
     >
-      <View
-        style={[
-          {
-            flex: 1,
-            flexDirection: "row",
-            alignItems: "center",
-          },
-        ]}
-      >
+      <View style={styles.row}>
         <Image
           source={require("../assets/magnifyingglass.png")}
-          style={{ width: 16, height: 16, marginLeft: 8, marginRight: 6 }}
+          style={styles.image}
           resizeMode="contain"
         />
-        <TextInput
-          style={{
-            fontSize: 17,
-          }}
-          placeholder="Search"
-        />
+        <TextInput style={styles.textInput} placeholder="Search" />
       </View>
     </View>
   );
 };
+
+const styles = StyleSheet.create({
+  textInput: {
+    fontSize: 17,
+  },
+  row: {
+    flex: 1,
+    flexDirection: "row",
+    alignItems: "center",
+  },
+  container: {
+    flex: 1,
+    borderRadius: 10,
+    backgroundColor: "#E3E3E8",
+    marginHorizontal: 16,
+  },
+  image: { width: 16, height: 16, marginLeft: 8, marginRight: 6 },
+});
